重構記憶體管理，統一型態並現代化程式碼
將記憶體管理改為直接使用 malloc/free/realloc，移除原有 GUIAlloc.c/GUIRealloc.c，新增 GUI_X_MEM.c。統一 GUI_HMEM/GUI_HWIN 為 void*，調整 WM_MESSAGE 結構與相關 API 以支援指標傳遞。main.c 移除 auto，改用明確型態。同步優化結構體格式、型態定義與 include，提升可讀性與可維護性。

+++ GUI/WM/WM.h
@@ -185,7 +185,7 @@ typedef struct {
   WM_HWIN hWinSrc;      /* Source window  */
   union {
     const void* p;            /* Some messages need more info ... Pointer is declared "const" because some systems (M16C) have 4 byte const, byte 2 byte default ptrs */
-    int v;
+    intptr_t v;
     RGB_COLOR Color;
   } Data;
 } WM_MESSAGE;
